First sketch of a condition system
This implementation uses a thread_local predicate (signal/2) as the call
stack's repository of logic for converting a condition into a restart.
The clauses or signal/2 are managed via setup_call_cleanup/3.

This doesn't quite work correctly if a handle/3 goal leaves
choicepoints. That edge case can be addressed later.  Alternatively, I
could declare that Goal is executed as once(Goal) so the caller knows
that choicepoints are discarded.  I don't yet have enough experience
with this code to know which approach makes the most sense.